Update atoz.md (#299)
* Update atoz.md

Fix json section by removing json syntax highlighting. The `//` comments make it invalid and it looks weird. so i adjusted the note, tweaks the example to be valid without comments and one or two settings that were wrong and i think that will be acceptable.

* Update atoz.md

keep spaces instead of tabs